addition of point in polygon routines
This commit adds a set of new functions for determining
whether a point P is located in a polygon given by its
vertex points V. Both algorithms for the even-odd-rule and
for the winding number calculation are implemented.

The commit also fixes a few typos in the documentation
of a few other routines.